move discovery, scheme and informer to pkg/runtime
