+++ pkg/runtime/discovery/dynamic_discovery.go
@@ -17,7 +17,7 @@ import (
 	"k8s.io/client-go/tools/cache"
 	"k8s.io/klog/v2"
 
-	"github.com/clusterpedia-io/clusterpedia/pkg/discovery/controller"
+	"github.com/clusterpedia-io/clusterpedia/pkg/runtime/discovery/controller"
 )
 
 type DynamicDiscoveryInterface interface {
