+++ pkg/runtime/informer/pager/result_stream.go
